// web/src/generated/gm-timeline/place.ts
/* eslint-disable */
// AUTO-GENERATED FILE. DO NOT EDIT.
// Generated by web/scripts/generate-gm-timeline-types.mjs.


import type { CampaignId, ReadAloudSection, Timestamp } from './shared-defs';

export type Place = {
  campaignId: CampaignId;
  description: string;
  headline: string;
  hookIds: Array<string>;
  id: string;
  notes?: string;
  readAloudSections?: Array<ReadAloudSection>;
  tags?: Array<string>;
  updatedAt?: Timestamp;
};