// web/src/generated/gm-timeline/place-index.ts
/* eslint-disable */
// AUTO-GENERATED FILE. DO NOT EDIT.
// Generated by web/scripts/generate-gm-timeline-types.mjs.


export type PlaceIndexIndexItem = {
  description?: string;
  file: string;
  headline: string;
  id: string;
  tags?: Array<string>;
};

export type PlaceIndex = {
  generatedAt?: string;
  items: Array<PlaceIndexIndexItem>;
};